+++ inlang/packages/paraglide/paraglide-js/src/compiler/compile.test.ts
@@ -184,7 +184,14 @@ test("multiple compile calls do not interfere with each other", async () => {
 
 	const runtime = (await import(
 		"data:text/javascript;base64," +
-			Buffer.from(runtimeFile, "utf-8").toString("base64")
+			Buffer.from(
+				// replace the
+				runtimeFile.replace(
+					`import "@inlang/paraglide-js/urlpattern-polyfill";`,
+					""
+				),
+				"utf-8"
+			).toString("base64")
 	)) as Runtime;
 
 	// expecting the second compile step to overwrite the first
@@ -286,6 +293,6 @@ test("default compiler options should include variable and baseLocale to ensure 
 	// paraglide js is the right tool for them, they can then define their own strategy.
 
 	expect(defaultCompilerOptions.strategy).toEqual(
-		expect.arrayContaining(["variable", "baseLocale"])
+		expect.arrayContaining(["globalVariable", "baseLocale"])
 	);
 });