+++ inlang/packages/paraglide/paraglide-js/src/compiler/runtime/get-locale.js
@@ -1,7 +1,13 @@
 import { assertIsLocale } from "./assert-is-locale.js";
-import { baseLocale, strategy } from "./variables.js";
+import {
+	baseLocale,
+	strategy,
+	TREE_SHAKE_COOKIE_STRATEGY_USED,
+	TREE_SHAKE_GLOBAL_VARIABLE_STRATEGY_USED,
+	TREE_SHAKE_URL_STRATEGY_USED,
+} from "./variables.js";
 import { extractLocaleFromCookie } from "./extract-locale-from-cookie.js";
-import { extractLocaleFromPathname } from "./extract-locale-from-pathname.js";
+import { extractLocaleFromUrl } from "./extract-locale-from-url.js";
 
 /**
  * This is a fallback to get started with a custom
@@ -31,20 +37,21 @@ export let getLocale = () => {
 	let locale;
 
 	for (const strat of strategy) {
-		if (strat === "cookie") {
+		if (TREE_SHAKE_COOKIE_STRATEGY_USED && strat === "cookie") {
 			locale = extractLocaleFromCookie();
-		}
-		if (strat === "baseLocale") {
+		} else if (strat === "baseLocale") {
 			locale = baseLocale;
-		}
-		if (
-			strat === "pathname" &&
-			typeof window !== "undefined" &&
-			window.location?.pathname
+		} else if (
+			TREE_SHAKE_URL_STRATEGY_USED &&
+			strat === "url" &&
+			typeof window !== "undefined"
 		) {
-			locale = extractLocaleFromPathname(window.location.pathname);
-		}
-		if (strat === "variable" && _locale !== undefined) {
+			locale = extractLocaleFromUrl(window.location.href);
+		} else if (
+			TREE_SHAKE_GLOBAL_VARIABLE_STRATEGY_USED &&
+			strat === "globalVariable" &&
+			_locale !== undefined
+		) {
 			locale = _locale;
 		}
 		// check if match, else continue loop
@@ -54,4 +61,23 @@ export let getLocale = () => {
 	}
 
 	throw new Error("No locale found. There is an error in your strategy.");
+};
+
+/**
+ * Define the \`getLocale()\` function.
+ *
+ * Use this function to define how the locale is resolved. For example,
+ * you can resolve the locale from the browser's preferred language,
+ * a cookie, env variable, or a user's preference.
+ *
+ * @example
+ *   defineGetLocale(() => {
+ *     // resolve the locale from a cookie. fallback to the base locale.
+ *     return Cookies.get('locale') ?? baseLocale
+ *   }
+ *
+ * @type {(fn: () => Locale) => void}
+ */
+export const defineGetLocale = (fn) => {
+	getLocale = fn;
 };