+++ inlang/packages/paraglide/paraglide-js/src/compiler/runtime/get-locale.test.ts
@@ -21,27 +21,12 @@ test("matching by strategy works", async () => {
 	expect(locale).toBe(baseLocale);
 });
 
-test("doesn't throw when using pathname strategy on the server", async () => {
-	const baseLocale = "en";
-
-	const runtime = await createRuntimeForTesting({
-		baseLocale,
-		locales: ["en", "de"],
-		compilerOptions: {
-			strategy: ["pathname", "variable", "baseLocale"],
-		},
-	});
-
-	expect(() => runtime.getLocale()).not.toThrow();
-	expect(runtime.getLocale()).toBe(baseLocale);
-});
-
 test("throws if variable is used without baseLocale as fallback strategy", async () => {
 	const runtime = await createRuntimeForTesting({
 		baseLocale: "en",
 		locales: ["en", "de"],
 		compilerOptions: {
-			strategy: ["variable"],
+			strategy: ["globalVariable"],
 		},
 	});
 
@@ -50,4 +35,54 @@ test("throws if variable is used without baseLocale as fallback strategy", async
 	runtime.setLocale("de");
 
 	expect(runtime.getLocale()).toBe("de");
+});
+
+test("retrieves the locale for a url pattern", async () => {
+	const runtime = await createRuntimeForTesting({
+		baseLocale: "en",
+		locales: ["en", "de"],
+		compilerOptions: {
+			strategy: ["url"],
+			urlPatterns: [
+				{
+					pattern: "https://example.:tld/:path*",
+					deLocalizedNamedGroups: { tld: "com" },
+					localizedNamedGroups: {
+						en: { tld: "com" },
+						de: { tld: "de" },
+					},
+				},
+			],
+		},
+	});
+
+	globalThis.window = { location: { href: "https://example.com/page" } } as any;
+
+	expect(runtime.getLocale()).toBe("en");
+
+	globalThis.window = { location: { href: "https://example.de/page" } } as any;
+
+	expect(runtime.getLocale()).toBe("de");
+});
+
+test("url pattern strategy doesn't throw during SSR", async () => {
+	const runtime = await createRuntimeForTesting({
+		baseLocale: "en",
+		locales: ["en", "de"],
+		compilerOptions: {
+			strategy: ["url", "baseLocale"],
+			urlPatterns: [
+				{
+					pattern: "https://example.:tld/:path*",
+					deLocalizedNamedGroups: { tld: "com" },
+					localizedNamedGroups: {
+						en: { tld: "com" },
+						de: { tld: "de" },
+					},
+				},
+			],
+		},
+	});
+
+	expect(() => runtime.getLocale()).not.toThrow();
 });