+++ inlang/packages/paraglide/paraglide-js/src/compiler/runtime/variables.js
@@ -6,7 +6,7 @@
  *     // do something
  *   }
  */
-export const baseLocale = "<base-locale>";
+export const baseLocale = "en";
 
 /**
  * The project's locales that have been specified in the settings.
@@ -16,15 +16,36 @@ export const baseLocale = "<base-locale>";
  *     throw new Error('Locale is not available');
  *   }
  */
-export const locales = /** @type {const} */ (["<base-locale>"]);
+export const locales = /** @type {const} */ (["en", "de"]);
 
 /** @type {string} */
 export const cookieName = "<cookie-name>";
 
 /**
- * @type {Array<"cookie" | "baseLocale" | "pathname" | "variable">}
+ * @type {Array<"cookie" | "baseLocale" | "globalVariable" | "url">}
  */
-export const strategy = ["variable"];
+export const strategy = ["globalVariable"];
 
-/** @type {boolean} */
-export const pathnamePrefixDefaultLocale = false;
+/**
+ * The used URL patterns.
+ *
+ * @type {Array<{ pattern: string, deLocalizedNamedGroups: Record<string, string | null>, localizedNamedGroups: Record<string, Record<string, string | null>> }>}
+ */
+export const urlPatterns = [];
+
+/**
+ * Whether the pathnames are the default pathnames.
+ *
+ * If the default pathnames are used, the matching is done
+ * statically with no runtime matching logic which reduces
+ * the bundlesize.
+ *
+ * @type {boolean}
+ */
+export const TREE_SHAKE_IS_DEFAULT_PATHNAMES = false;
+
+export const TREE_SHAKE_COOKIE_STRATEGY_USED = false;
+
+export const TREE_SHAKE_URL_STRATEGY_USED = false;
+
+export const TREE_SHAKE_GLOBAL_VARIABLE_STRATEGY_USED = false;