+++ inlang/packages/paraglide/paraglide-js/CHANGELOG.md
@@ -1,13 +1,51 @@
 # @inlang/paraglide-js
 
+## 2.0.0-beta.18
+
+Added URLPatterns as a replacement for the beta 17 pathnames API. 
+
+The URLPattern API is extremly powerful. You can express base paths, translated pathnames, domain based localization, and even multi-tenancy. 
+
+Read the docs [here](https://inlang.com/m/gerre34r/library-inlang-paraglideJs/strategy#url) and make PRs to improve the documentation.
+
+```diff
+await compile({
+-  strategy: ["pathname"],
++  strategy: ["url"],
+})
+```
+
+The `localizePath()` API had to be replaced by a new `localizeHref()` API. Please give feedback on the new API in [#380](https://github.com/opral/inlang-paraglide-js/issues/380)
+
+```diff
+- <a href={localizePath("/about")}>About</a>
++ <a href={localizeHref("/about")}>About</a>
+```
+
 ## 2.0.0-beta.17
 
-Add support for `pathnamePrefixBaseLocale` https://github.com/opral/inlang-paraglide-js/issues/358.
+Add support for `pathnames` API https://github.com/opral/inlang-paraglide-js/issues/359
+
+You can now create whatever pathname pattern you want. The syntax is provided by https://github.com/pillarjs/path-to-regexp.
 
 ```diff
 await compile({
   strategy: ["pathname", "cookie", "baseLocale"],
-+ pathnamePrefixBaseLocale: true
++ pathnames: {
++   "{*path}": {
++     "de": "/de{/*path}",
++     "en": "/en{/*path}",
+}
+}
+})
+```
+
+Add support for `pathnameBase` https://github.com/opral/inlang-paraglide-js/issues/362
+
+```diff
+await compile({
+  strategy: ["pathname", "cookie", "baseLocale"],
++ pathnameBase: "/base"
 })
 
 ```
@@ -30,7 +68,7 @@ Fixes windows path normalization https://github.com/opral/monorepo/pull/3374
 
 ## 2.0.0-beta.13
 
-- flat compiler options 
+- flat compiler options
 
 ```diff
 await compile({
@@ -51,7 +89,7 @@ await compile({
 
 ## 2.0.0-beta.11
 
-- improve: compiler awaits ongoing compilations before starting a new one 
+- improve: compiler awaits ongoing compilations before starting a new one
 - update dependency that fixes https://github.com/opral/inlang-paraglide-js/issues/320
 
 ## 2.0.0-beta.10
@@ -60,7 +98,7 @@ await compile({
 
 ## 2.0.0-beta.9
 
-- Expose compiler on the index 
+- Expose compiler on the index
 
 ```diff
 -import { compile } from "@inlang/paraglide-js/compiler";
@@ -89,13 +127,13 @@ await compile({
 
 ### Patch Changes
 
-- remove `fast-glob` as dependency in favor of node's built-in `fs.glob` (a new API in node 22) 
+- remove `fast-glob` as dependency in favor of node's built-in `fs.glob` (a new API in node 22)
 
 ## 2.0.0
 
 ### Major changes
 
-- Upgrade to @inlang/sdk v2 
+- Upgrade to @inlang/sdk v2
 - Support for variants (pluralization, gendering, A/B test, etc.)
 
 ### Minor Changes